requirements/common.txt
django
jingo
south
django-staticfiles-jquery
django-staticfiles-bootstrap
django-browserid
django-braces
